+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## [1.2.5] - 2025-04-15
 ## add:
 - fpctoolkit.project.rebuild can be called with a shortcut (Shift+F9) - does not need a need to be called from
-- editor commands (fpctoolkit.editor.): base64encode, base64decode, hexencode, hexdecode, urlencode, urldecode
+- editor commands (fpctoolkit.editor.): base64encode, base64decode, hexencode, hexdecode, urlencode, urldecode, quoted-printable decode
 - ctrl+shift+g Generate UUID
 ## fix:
 - activate project - reloadTasks no longer work (used async call), client reloaded now